guns refactor
