+++ core/main.ts
@@ -108,6 +108,7 @@ export * from "./src/ecs/actions/movement/Move"
 export * from "./src/ecs/actions/movement/WASDInputMap"
 export * from "./src/ecs/actions/movement/VehicleMovement"
 export * from "./src/ecs/entities/items/Tools"
+export * from "./src/ecs/entities/items/Guns"
 export * from "./src/ecs/components/Controlling"
 export * from "./src/sound/SoundManager"
 export * from "./src/utils/MathUtils"
